changed method int DataStore.size() to long getNumberOfRecords(). Added more javadoc
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1089 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ test/org/jcvi/common/core/seq/read/trace/pyro/sff/TestReadExampleSffFile.java
@@ -53,7 +53,7 @@ public abstract class  TestReadExampleSffFile extends AbstractTestExampleSffFile
         assertEquals(FF585OX02GMGGN, getFlowgram("FF585OX02GMGGN"));
         assertEquals(FF585OX02FHO5X,getFlowgram("FF585OX02FHO5X"));
     }
-    protected abstract int getNumberOfFlowgrams() throws Exception;
+    protected abstract long getNumberOfFlowgrams() throws Exception;
    protected abstract Flowgram getFlowgram(String id) throws Exception;
    protected abstract void parseSff(File f) throws Exception;
 }
